+++ src/UserInterface.hpp
@@ -22,7 +22,7 @@ public:
     void towers_to_select(int &_width, int &_height, std::unordered_map<std::string, GLuint> textures, ItdTower ItdTower, float viewSize);
     void show_level(int level);
     void load_infos_targets(std::vector<Target> targets, int app_current_monster_index);
-    std::vector<std::pair<std::pair<int, int>, std::pair<int, int>>> get_tower_positions;
+    std::vector<std::pair<std::pair<float, float>, std::pair<float, float>>> get_tower_positions;
 
 private:
     Map map;
